AKS/Independent Solution/src/requirement.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-storage-blob
azure-identity
pandas
scikit-learn==0.22.2.post1